refactor(objectives): move Evidence routing contract into analysis
Why:
- Evidence routing already owned Source selection and queue order, but its model judgment was still split across generic extractor, prompt, and schema modules.

What:
- Co-locate the bounded routing prompt, response contract, role normalization, and model call with the existing analysis stage.

Impact:
- Source identities, route decisions, deterministic fallback, extraction order, persistence, and public contracts remain unchanged.

Verification:
- All 138 routing, workflow, and extractor tests and 6 collection-build integration tests passed.

Refs:
- ISSUE: #330
