+++ backend/tests/support/fake_domain_model_extractor.py
@@ -7,6 +7,10 @@ from application.core.document_profiles.schemas import StructuredDocumentProfile
 from application.core.objectives.analysis.source_screening import (
     StructuredPaperFrameBatch,
 )
+from application.core.objectives.analysis.evidence_routing import (
+    StructuredEvidenceSelection,
+    StructuredEvidenceSelections,
+)
 from application.core.objectives.discovery.axis_equivalence import (
     StructuredAxisCanonicalizationPlan,
 )
@@ -17,8 +21,6 @@ from application.core.objectives.discovery.study_window import StructuredPaperSk
 from application.core.objectives.schemas import (
     StructuredEvidenceExtraction,
     StructuredEvidenceExtractions,
-    StructuredEvidenceSelection,
-    StructuredEvidenceSelections,
 )
 from application.core.paper_facts.schemas import (
     MeasurementValuePayload,
@@ -352,7 +354,7 @@ class FakeDomainModelExtractor:
             excluded_source_unit_ids=excluded_source_unit_ids,
         )
 
-    def select_objective_evidence(
+    def route_source(
         self,
         payload: dict[str, Any],
     ) -> StructuredEvidenceSelections:
